Refactor portfolio allocation visuals
