+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -12,3 +12,4 @@
 | LWB-R008 | Projected cashflow chart component extraction | Hardened | After the chart view-model extraction, `portfolio-chart-panels.tsx` still owned four exported chart components, keeping unrelated holdings, cashflow, activity, and income rendering coupled in one large module. | Moved `PortfolioProjectedCashflowPanel` into `portfolio-projected-cashflow-panel.tsx` and kept the original `portfolio-chart-panels.tsx` named export as a compatibility surface for existing dynamic imports and callers. | Continue extracting `PortfolioTopHoldingsPanel`, `PortfolioActivityPanel`, and `PortfolioIncomePanel` into separate files, then collapse `portfolio-chart-panels.tsx` into a small export barrel. | `src/apps/portfolio/components/portfolio-projected-cashflow-panel.tsx`, `src/apps/portfolio/components/portfolio-chart-panels.tsx`, validation via `npm test -- --run tests/unit/portfolio-chart-view-model.test.ts tests/unit/portfolio-chart-panels.test.tsx` |
 | LWB-R009 | Portfolio chart component barrel split | Hardened | `portfolio-chart-panels.tsx` still coupled the remaining chart implementations after the projected cashflow extraction, so unrelated holdings, activity, income, and cashflow chart rendering remained grouped behind a large mixed module. | Split `PortfolioTopHoldingsPanel`, `PortfolioActivityPanel`, and `PortfolioIncomePanel` into dedicated component files and reduced `portfolio-chart-panels.tsx` to a stable named-export compatibility barrel. Existing dynamic imports continue to resolve through the same module path while each chart can now evolve independently. | Consider replacing compatibility-barrel dynamic imports with direct component imports once downstream churn is acceptable. The next monolith target is `portfolio-allocation-panel.tsx`, which still combines filter state, visualization, grid, and drawer concerns. | `src/apps/portfolio/components/portfolio-top-holdings-panel.tsx`, `src/apps/portfolio/components/portfolio-activity-panel.tsx`, `src/apps/portfolio/components/portfolio-income-panel.tsx`, `src/apps/portfolio/components/portfolio-chart-panels.tsx`; validation via `npm test -- --run tests/unit/portfolio-chart-view-model.test.ts tests/unit/portfolio-chart-panels.test.tsx tests/unit/portfolio-paired-analytics-section.test.tsx`; import scan via `rg -n "portfolio-chart-panels" src/apps/portfolio tests` |
 | LWB-R010 | Portfolio allocation pure view-model extraction | Hardened | `portfolio-allocation-panel.tsx` mixed rendering, Gateway request orchestration, allocation constants, look-through support rules, dimension label formatting, and donut SVG geometry. The pure logic was hard to test without rendering the full panel. | Added `portfolio-allocation-view-model.ts` for allocation dimensions, chart types, palette, label formatting, look-through normalization/support checks, and deterministic donut arc geometry; updated the panel to consume these helpers; added unit tests for constants, support posture, contract-safe fixtures, and arc output while preserving existing panel tests. | Continue by extracting the allocation visual subcomponents or the look-through request state machine into focused modules. Keep Gateway calls inside the panel/controller layer until a reusable allocation data hook is justified. | `src/apps/portfolio/portfolio-allocation-view-model.ts`, `src/apps/portfolio/components/portfolio-allocation-panel.tsx`, `tests/unit/portfolio-allocation-view-model.test.ts`; validation via `npm test -- --run tests/unit/portfolio-allocation-view-model.test.ts tests/unit/portfolio-allocation-panel.test.tsx`, `npm run typecheck`, `npm run lint`, `git diff --check` |
+| LWB-R011 | Portfolio allocation visual component extraction | Hardened | `portfolio-allocation-panel.tsx` still mixed Gateway-backed look-through orchestration with donut, bar, table, ranked-list, and empty-state rendering. The file had become a broad allocation controller plus chart gallery, making future visual changes harder to review without touching request state. | Moved allocation visual subcomponents into `portfolio-allocation-visuals.tsx` and left the panel focused on allocation state, Gateway look-through hydration, dimension/chart selection, and parent drill-down callbacks. Existing panel tests continue to cover keyboard selection, compact layout, empty state, look-through enablement, and Gateway fallback behavior. | Extract the look-through request state machine only if it starts being reused or if the panel controller grows again. Continue scanning `portfolio-workspace.tsx`, holdings grid, and transactions grid for high-value modularity improvements. | `src/apps/portfolio/components/portfolio-allocation-panel.tsx`, `src/apps/portfolio/components/portfolio-allocation-visuals.tsx`, `tests/unit/portfolio-allocation-panel.test.tsx`; validation via `npm test -- --run tests/unit/portfolio-allocation-view-model.test.ts tests/unit/portfolio-allocation-panel.test.tsx` |
